fix: 모바일과 데스크탑 환경을 분리하여 timecapsop모달 반응형 처리
- 데스트탑일 경우 일반 모달, 모바일일 경우 bottomsheet 모달 스타일로 수정
