+++ src/components/common/HomePopup/index.tsx
@@ -109,16 +109,15 @@ export const HomePopup = () => {
         </StBackground>
       )}
 
-      {isOpenResolutionModal && (
-        <TimecapsopSubmitModal
-          onClose={onCloseResolutionModal}
-          userName={name ?? '나'}
-          onSuccess={() => {
-            onNewRegistration();
-            onOpenPlaygroundGuideModal();
-          }}
-        />
-      )}
+      <TimecapsopSubmitModal
+        onClose={onCloseResolutionModal}
+        userName={name ?? '나'}
+        onSuccess={() => {
+          onNewRegistration();
+          onOpenPlaygroundGuideModal();
+        }}
+        isOpen={isOpenResolutionModal}
+      />
       {isOpenPlaygroundGuideModal && (
         <PlaygroundGuideModal isAlreadyRegistration={isAlreadyRegistration} onClose={onClosePlaygroundGuideModal} />
       )}
