+++ src/pages/members/complete.tsx
@@ -123,13 +123,12 @@ const CompletePage: FC = () => {
           </Responsive>
         </StyledCompletePage>
       )}
-      {isOpenResolutionModal && (
-        <TimecapsopSubmitModal
-          userName={name ?? '나'}
-          onClose={onCloseResolutionModal}
-          onSuccess={onOpenPlaygroundGuideModal}
-        />
-      )}
+      <TimecapsopSubmitModal
+        userName={name ?? '나'}
+        onClose={onCloseResolutionModal}
+        onSuccess={onOpenPlaygroundGuideModal}
+        isOpen={isOpenResolutionModal}
+      />
       {isOpenPlaygroundGuideModal && (
         <PlaygroundGuideModal
           onClose={() => {
